file: UITableView, UICollectionView의 익스텐션을 Extension 폴더로 이동
- dequeueReusableCell 메서드는 비지니스 로직이 없는 순수 유틸로 기능하므로 Extension file로 분리함
