+++ OpenMarket/OpenMarket/Controller/ProductCollectinoViewController.swift
@@ -107,22 +107,3 @@ final class ProductCollectinoViewController: UICollectionViewController {
         }
     }
 }
-
-// MARK: - UICollectionView Extension
-
-private extension UICollectionView {
-    
-    func dequeueReusableCell<T: UICollectionViewCell>(
-        withClass name: T.Type,
-        for indexPath: IndexPath
-    ) -> T {
-        guard let cell = dequeueReusableCell(
-            withReuseIdentifier: String(describing: name),
-            for: indexPath
-        ) as? T else {
-            fatalError(
-                "Couldn't find UICollectionViewCell for \(String(describing: name)), make sure the cell is registered with collection view")
-        }
-        return cell
-    }
-}
